+++ addOns/ascanrules/CHANGELOG.md
@@ -4,8 +4,13 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
+### Added
+- Checks for cloud metadata from IBM and OpenStack.
+- Evidence for cloud metadata.
 ### Changed
 - Update dependency.
+### Fixed
+- Cloud metadata false positives by making the evidence checks more specific.
 
 ## [79] - 2025-12-30
 ### Changed